refactor Board module
