+++ src/App.tsx
@@ -1,6 +1,5 @@
 // src/App.tsx
-import React from 'react';
-import Board from './Board';
+import Board from './Board/Board';
 function App() {
   return (
     <div className="App">
