+++ src/Board/Timer.tsx
